feat(api): unify postulation and classification
